LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/object.https.sub-expected.txt



PASS Same-Origin object
FAIL Same-Site object assert_not_equals: got disallowed value "No header has been recorded"
FAIL Cross-Site object assert_not_equals: got disallowed value "No header has been recorded"
FAIL Same-Origin object: sec-fetch-dest assert_equals: expected "object" but got ""
FAIL Same-Origin object: sec-fetch-mode assert_equals: expected "navigate" but got ""
FAIL Same-Origin object: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS Same-Origin object: sec-fetch-user
